Add Investment type/Lottery

+++ contracts/libs/TokenFormat.sol
@@ -1,6 +1,9 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 pragma solidity ^0.7.0;
 
+/**
+ * @title The Token Format library
+ */
 library TokenFormat {
     // Use a split bit implementation.
     // Store the generation in the upper 128 bits..
@@ -8,17 +11,23 @@ library TokenFormat {
     uint256 private constant _LOAN_ID_MASK = uint128(~0);
 
     /**
-     * @dev Format tokenId into generation and index
+     * @notice Format tokenId into generation and index
+     * @param tokenId The Id of the token
+     * @return generation
+     * @return loanId
      */
-    function formatTokenId(uint tokenId) internal pure returns(uint generation, uint loanId) {
+    function formatTokenId(uint256 tokenId) internal pure returns (uint256 generation, uint256 loanId) {
         generation = tokenId >> 128;
         loanId = tokenId & _LOAN_ID_MASK;
     }
 
     /**
-     * @dev get tokenId from generation and loanId
+     * @notice get tokenId from generation and loanId
+     * @param gen the generation
+     * @param loanId the loanID
+     * @return tokenId the token id
      */
-    function getTokenId(uint gen, uint loanId) internal pure returns(uint tokenId) {
+    function getTokenId(uint256 gen, uint256 loanId) internal pure returns (uint256 tokenId) {
         return (gen << 128) | loanId;
     }
 }
